Avoid /login authentication when using Gerrit HTTP password tokens.
This change introduces a new contructor in the GerritAuthData class.
The new constructor takes an additional "httpPassword" argument,
which is used to identify whether the password being sepcified is
a special Gerrit HTTP password token (typically a password which
can be generated for each user on their Settings -> HTTP Password
page).  When set to false (the default), the password is assumed
to be a valid password for whatever authentication source Gerrit
is configured to use for normal user authentication.  When set to
true, the password is assumed to be a HTTP Password maintained
internally by Gerrit.

This fix is intended to resolve a problem where specifying an
HTTP Password for REST API calls will attempt to authenticate
against the /login page.  When Gerrit is configured to authenticate
against LDAP, this will cause authentication failure and typically
lock the LDAP account after a number of invalid attempts.