Check if session is expired when using GerritAuth cookie authentication
When the session is expired / invalid, a request to fetch changes will
not fail but it won't return the expected results. When using the cookie
authentication method, we issue a check request before every real request.
This is not optimal for performance reasons, but better than an
unexpected (wrong) result.