Avoid spamming Gerrit when using GitHub/OAuth authentication
When Gerrit is configured with a GitHub/OAuth auth type,
the form authentication is technically available but not
in a way that can be automatically used by the REST Client.

GitHub may require additional authentication verification
methods such as 2FA that are incompatible with a batch use
by the Gerrit REST client library.

Detect the OAuth handshake by looking at the redirection
path containing the github plugin scope selection and
avoid attempting the form authentication mechanism again.

Closes #107